requirements.txt
colorlog
graphviz
huggingface_hub
inflection
jinja2
jsonpointer
kubernetes
langchain
langchain-openai
matplotlib
networkx
# openark
opentelemetry-distro
opentelemetry-exporter-otlp
ortools
pandas
polars[pyarrow]
pycairo
pydantic
PyGObject
python-dotenv
pyvis
pyyaml
scipy
streamlit
streamlit-aggrid
streamlit-folium
streamlit-navigation-bar
